+++ src/main/java/com/irum/orderservice/domain/order/service/CustomerOrderService.java
@@ -1,11 +1,11 @@
 package com.irum.orderservice.domain.order.service;
 
 import com.irum.global.advice.exception.CommonException;
-import com.irum.orderservice.domain.client.payment.PaymentClient;
-import com.irum.orderservice.domain.client.payment.dto.emuns.PaymentCorp;
-import com.irum.orderservice.domain.client.payment.dto.response.PaymentResponse;
-import com.irum.orderservice.domain.client.product.ProductClient;
-import com.irum.orderservice.domain.client.product.dto.response.ProductInternalResponse;
+import com.irum.orderservice.openfeign.payment.PaymentClient;
+import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentCorp;
+import com.irum.orderservice.openfeign.payment.dto.response.PaymentResponse;
+import com.irum.orderservice.openfeign.product.ProductClient;
+import com.irum.orderservice.openfeign.product.dto.response.ProductInternalResponse;
 import com.irum.orderservice.domain.coupon.service.AppliedCouponService;
 import com.irum.orderservice.domain.coupon.service.CouponService;
 import com.irum.orderservice.domain.deliveryaddress.domain.entity.DeliveryAddress;
