+++ src/main/java/com/irum/orderservice/openfeign/payment/PaymentClient.java
@@ -1,10 +1,10 @@
-package com.irum.orderservice.domain.client.payment;
+package com.irum.orderservice.openfeign.payment;
 
-import com.irum.orderservice.domain.client.payment.api.PaymentAPI;
-import com.irum.orderservice.domain.client.payment.dto.emuns.PaymentCorp;
-import com.irum.orderservice.domain.client.payment.dto.request.CreatePaymentRequest;
-import com.irum.orderservice.domain.client.payment.dto.request.UpdatePaymentStatusRequest;
-import com.irum.orderservice.domain.client.payment.dto.response.PaymentResponse;
+import com.irum.orderservice.openfeign.payment.api.PaymentAPI;
+import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentCorp;
+import com.irum.orderservice.openfeign.payment.dto.request.CreatePaymentRequest;
+import com.irum.orderservice.openfeign.payment.dto.request.UpdatePaymentStatusRequest;
+import com.irum.orderservice.openfeign.payment.dto.response.PaymentResponse;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
