+++ src/main/java/com/irum/orderservice/openfeign/payment/dto/emuns/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.irum.orderservice.domain.client.payment.dto.emuns;
+package com.irum.orderservice.openfeign.payment.dto.emuns;
 
 public enum PaymentStatus {
     FAILED,
