+++ src/main/java/com/irum/orderservice/openfeign/payment/dto/response/PaymentResponse.java
@@ -1,0 +1,10 @@
+package com.irum.orderservice.openfeign.payment.dto.response;
+
+import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentMethod;
+import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentStatus;
+
+public record PaymentResponse(
+        PaymentStatus paymentStatus,
+        PaymentMethod paymentMethod,
+        int totalDiscountAmount,
+        int amount) {}
