+++ src/main/java/com/irum/orderservice/openfeign/product/api/ProductAPI.java
@@ -1,9 +1,9 @@
-package com.irum.orderservice.domain.client.product.api;
+package com.irum.orderservice.openfeign.product.api;
 
-import com.irum.orderservice.domain.client.config.FeignConfig;
-import com.irum.orderservice.domain.client.product.dto.request.ProductInternalRequest;
-import com.irum.orderservice.domain.client.product.dto.request.RollbackStockRequest;
-import com.irum.orderservice.domain.client.product.dto.response.ProductInternalResponse;
+import com.irum.orderservice.openfeign.config.FeignConfig;
+import com.irum.orderservice.openfeign.product.dto.request.ProductInternalRequest;
+import com.irum.orderservice.openfeign.product.dto.request.RollbackStockRequest;
+import com.irum.orderservice.openfeign.product.dto.response.ProductInternalResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
