+++ src/main/java/com/irum/orderservice/openfeign/product/dto/request/ProductInternalRequest.java
@@ -1,11 +1,11 @@
-package com.irum.orderservice.domain.client.product.dto.request;
+package com.irum.orderservice.openfeign.product.dto.request;
 
 import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record RollbackStockRequest(List<OptionValueRequest> optionValueList) {
+public record ProductInternalRequest(List<OptionValueRequest> optionValueList, UUID storeId) {
     @Builder
     public record OptionValueRequest(UUID optionValueId, int quantity) {}
 }
