+++ src/main/java/com/irum/orderservice/openfeign/product/dto/request/RollbackStockRequest.java
@@ -1,11 +1,11 @@
-package com.irum.orderservice.domain.client.product.dto.request;
+package com.irum.orderservice.openfeign.product.dto.request;
 
 import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record ProductInternalRequest(List<OptionValueRequest> optionValueList, UUID storeId) {
+public record RollbackStockRequest(List<OptionValueRequest> optionValueList) {
     @Builder
     public record OptionValueRequest(UUID optionValueId, int quantity) {}
 }
